vk-turf-update: bumped Turf version to `0.5.0`. Removed CoordinateBounds in favor of Turf.BoundingBox.
